analyzer: split expression.{h,cpp} somewhat.
