+++ include/reaver/vapor/analyzer/expressions/expression.h
@@ -121,136 +121,6 @@ inline namespace _v1
         std::unique_ptr<variable> _variable;
     };
 
-    class expression_list : public expression
-    {
-    private:
-        virtual future<> _analyze(analysis_context &) override;
-        virtual std::unique_ptr<expression> _clone_expr_with_replacement(replacements &) const override;
-        virtual future<expression *> _simplify_expr(simplification_context &) override;
-
-        virtual statement_ir _codegen_ir(ir_generation_context & ctx) const override
-        {
-            return mbind(value, [&](auto && expr) { return expr->codegen_ir(ctx); });
-        }
-
-    public:
-        expression_list() = default;
-
-        virtual void print(std::ostream & os, std::size_t indent) const override;
-
-        virtual variable * get_variable() const override
-        {
-            return value.back()->get_variable();
-        }
-
-        range_type range;
-        std::vector<std::unique_ptr<expression>> value;
-    };
-
-    class variable_ref_expression : public expression
-    {
-    public:
-        variable_ref_expression(variable * var) : _referenced(var)
-        {
-        }
-
-        virtual variable * get_variable() const final override
-        {
-            return _referenced;
-        }
-
-        virtual void print(std::ostream & os, std::size_t indent) const override;
-
-    private:
-        virtual future<> _analyze(analysis_context &) override
-        {
-            return make_ready_future();
-        }
-
-        virtual std::unique_ptr<expression> _clone_expr_with_replacement(replacements & repl) const override
-        {
-            auto it = repl.variables.find(_referenced);
-            if (it == repl.variables.end())
-            {
-                return std::make_unique<variable_ref_expression>(_referenced);
-            }
-
-            return std::make_unique<variable_ref_expression>(it->second);
-        }
-
-        virtual future<expression *> _simplify_expr(simplification_context & ctx) override
-        {
-            return _referenced->simplify(ctx).then([&](auto && simplified) -> expression * {
-                _referenced = simplified;
-                return this;
-            });
-        }
-
-        virtual statement_ir _codegen_ir(ir_generation_context & ctx) const override
-        {
-            return { codegen::ir::instruction{
-                none, none, { boost::typeindex::type_id<codegen::ir::pass_value_instruction>() }, {}, get<codegen::ir::value>(_referenced->codegen_ir(ctx)) } };
-        }
-
-        variable * _referenced;
-    };
-
-    inline std::unique_ptr<expression> make_variable_ref_expression(variable * var)
-    {
-        return std::make_unique<variable_ref_expression>(var);
-    }
-
-    class variable_expression : public expression
-    {
-    public:
-        variable_expression(std::unique_ptr<variable> var)
-        {
-            _set_variable(std::move(var));
-        }
-
-        virtual void print(std::ostream & os, std::size_t indent) const override;
-
-    private:
-        virtual future<> _analyze(analysis_context &) override
-        {
-            return make_ready_future();
-        }
-
-        std::unique_ptr<expression> _clone_expr_with_replacement(replacements & repl) const override
-        {
-            auto it = repl.variables.find(get_variable());
-            if (it == repl.variables.end())
-            {
-                return std::make_unique<variable_expression>(get_variable()->clone_with_replacement(repl));
-            }
-
-            return make_variable_ref_expression(it->second);
-        }
-
-        virtual future<expression *> _simplify_expr(simplification_context & ctx) override
-        {
-            return get_variable()->simplify(ctx).then([&](auto && simplified) -> expression * {
-                this->_set_variable(simplified, ctx);
-                return this;
-            });
-        }
-
-        virtual statement_ir _codegen_ir(ir_generation_context & ctx) const override
-        {
-            return { codegen::ir::instruction{ none,
-                none,
-                { boost::typeindex::type_id<codegen::ir::pass_value_instruction>() },
-                {},
-                get<codegen::ir::value>(get_variable()->codegen_ir(ctx)) } };
-        }
-    };
-
-    inline std::unique_ptr<expression> make_variable_expression(std::unique_ptr<variable> var)
-    {
-        return std::make_unique<variable_expression>(std::move(var));
-    }
-
     std::unique_ptr<expression> preanalyze_expression(const parser::expression & expr, scope * lex_scope);
-    std::unique_ptr<expression> preanalyze_expression(const parser::expression_list & expr, scope * lex_scope);
 }
 }
